requirements.txt
python-telegram-bot
environs
nltk
